\name{socceR-package}
\alias{socceR-package}
\alias{socceR}
\docType{package}
\title{
\packageTitle{socceR}
}
\description{
\packageDescription{socceR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{socceR}
\packageIndices{socceR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{socceR}

Maintainer: \packageMaintainer{socceR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
